move actions to right folder and rename files
