+++ apps/web/src/app/actions/badges/badge_types/advent_badges.ts
@@ -1,7 +1,7 @@
 import type {AllBadgeObjs, BadgesFn} from "../_actions";
 import {prisma} from "@repo/db";
-import type {difficultyBadges} from "~/app/(profile)/[username]/_components/dashboard/badges/_difficulty_badges";
-import type {SolutionBadges} from "~/app/(profile)/[username]/_components/dashboard/badges/_shared_solutions_badges";
+import type {difficultyBadges} from "~/app/actions/badges/badge_types/difficulty_badges";
+import type {SolutionBadges} from "~/app/actions/badges/badge_types/shared_solutions_badges";
 
 export interface AdventChallenges {
   trackChallenges: { challenge: { submission: { isSuccessful: boolean }[] } }[];
