+++ apps/web/src/app/challenge/_components/comments/enhanced-user-badge.getProfileData.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import { prisma } from '@repo/db';
-import { getBadges } from '~/app/(profile)/[username]/_components/dashboard/_actions';
+import { getBadges } from '~/app/actions/badges/_actions';
 import { getTitles } from './enhanced-user-badge.getTitles';
 
 export async function getProfileData(username: string) {
