+++ apps/web/src/app/challenge/_components/comments/enhanced-user-badge.tsx
@@ -8,7 +8,7 @@ import { useQuery } from '@tanstack/react-query';
 import { getProfileData } from './enhanced-user-badge.getProfileData';
 import { useState } from 'react';
 import { SlugToBadgeIcon } from '~/app/(profile)/[username]/_components/dashboard/badges';
-import {AllBadges, type AOT2023} from '~/app/(profile)/[username]/_components/dashboard/_actions';
+import {AllBadges, type AOT2023} from '~/app/actions/badges/_actions';
 import { cn } from '@repo/ui/cn';
 import { type Role } from '@repo/db/types';
 import { Button } from '@repo/ui/components/button';
